+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [ADDED] - new instance methods `has()` & `remove()` of the `ResourceManager`
 * [ADDED] - new instance method `has()` of the `RemoteServiceManager`
 * [ADDED] - `services.<name>` configuration section can now contain additional axios constructor options which the SDK will be instantiated with at the time of `remoteServiceManager.buildRemoteService()` call
+* [FIXED] - `service` static event of `Service` constructor object should be emitted in the next event loop cycle to give plugins a chance to hook-up
 
 ## v1.2.0